+++ pcweb/pages/demos_on_landing_page/chatbox/chat_demo.py
@@ -1,0 +1,13 @@
+import reflex as rx
+
+def chat_box():
+    return rx.vstack(
+        rx.text("hello", color="red"),
+        rx.text("world", color="red"),
+        rx.text("hello", color="red"),
+        rx.text("world", color="red"),
+        rx.text("hello", color="red"),
+        rx.text("world", color="red"),
+        rx.text("hello", color="red"),
+        rx.text("world", color="red"),
+    )
