+++ pcweb/pages/demos_on_landing_page/image_gen/image_gen.py
@@ -1,0 +1,4 @@
+import reflex as rx
+
+def image_gen():
+    return
